engine: Add method for checking existence of network in cluster
The patch adds a method to check if a given network is assigned
to a given cluster. The method will be used in can-do-action
validation where the vnic profile set for a vnic should be
verified as a vnic of a network within the cluster of the vm or
of the template.

Change-Id: I95e6f23ebc17f2b06c06590d634e5fa113f23a29
Signed-off-by: Moti Asayag <masayag@redhat.com>