restapi: image entity and import from glance
This patch adds an initial implementation for the REST API commands to
import and export images from and to image storage domains (as OpenStack
Glance).

The API at the moment has to be intended as Technology Preview and it
might be subject to changes in the future.

Glance integration feature page:

 http://www.ovirt.org/Features/Glance_Integration

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=990962
Change-Id: Ief9104ff0eba175c90b42bc5651ada0f688eeb00
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>