packaging: spec: setup: do not leave empty directories on remove
when installing files via wildcards, for example:
%{_datadir}/*/*.py

rpm does not record the directory within his archive, so upon removal
this directory is left, while the files are removed.

we do want to avoid the need to specify each file within the spec file,
as source tree is organized properly, we can add the top level.

however, in future if we have the need to split the directory structure
into two packages, we will have to work a bit harder while maintaining
the spec file, we may consider to generate it(?).

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=988685
Change-Id: I63fe12c6f29ff82588dfabb5df93eaca12154c12
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>