core: set spice as default display type
When creating a VM, its default display type is taken from the template
it is based on. This patch sets the default display type of the Blank
template to SPICE so that SPICE will be set as the default display type
when creating VM which is based on the Blank template.

Change-Id: Icfc38462e04f4d7ebcd40f90819feec87d214791
Bug-Url: https://bugzilla.redhat.com/985313
Signed-off-by: Arik Hadas <ahadas@redhat.com>