core: handle 'image does not exists' error on remove hibernation vols
Introducing HibernationVolumesRemover class which removes hibernation
volumes with proper handling of 'image does not exists' error. This
error is not treated as an error, because the result is that the images
don't exists, as we wanted.

HibernationVolumesRemover class is being used by
RemoveVmHibernationVolumesCommand which is used by stop VM commands
(stop and shutdown) and run VM command (in resume from hibernation flow
in particular), eliminating the race described in bz 952603.

Change-Id: Ief52a20315b29660c7038eab1cc8dbbb4c4ded3e
Signed-off-by: Arik Hadas <ahadas@redhat.com>