restapi: revert 925aeee7 Add vnic custom properties
The custom properties per device will be managed on vnic profile
level instead for vm nic.

The motivation of reverting this patch is not to commit to custom
properties api on vnic level, unless a strong use-case for it
which isn't satisfied by the vnic profile is presented.

Therefore the commit which adds it should be reverted.

Change-Id: Ie893f50addf631e72dc52f1c740cf9bb41305dfd
Signed-off-by: Moti Asayag <masayag@redhat.com>