restapi: allow storage conn deletion with no host
Allow deletion of storage connection without specifying the host.
Before this change, deletion of storage connection required host id/name
in order to perform a disconnect operation as part of the connection
deletion.
This change allows deleting the connection from db without the need to
disconnect the host from it, since the connection could have been
created without connecting the host to it.
The option to delete connection and disconnect it via the host remains
supported as well.

Feature Page:
http://www.ovirt.org/Features/Manage_Storage_Connections

Change-Id: I7cd9dcea95028e12872beb2496c4338654cfbdb4
Signed-off-by: Alissa Bonas <abonas@redhat.com>