core: improve active iso domain lookup
The lookup for active iso domain at
IsoDomainListSyncronizer#findActiveISODomain access the DB twice: at
first to fetch all the domains in the given storage pool and then it
access the DB again to fetch a given domain within the same storate
pool. the second access seems to be redundant since the fetched domain
already exists in the list that was return from the first query, thus
the second query is removed.

Change-Id: I8859a85107f8c2e658c23d6a04b69d0201d086d5
Signed-off-by: Arik Hadas <ahadas@redhat.com>