core: Take into account command transactivity scope when...
Take into account command transactivity scope when creating new
task duing persistence of place holder.
If the command is transactive - create the place holder in a new
transaction, so the record will be recognized when there will
be association between VDSM task ID and db task ID.
This will solve the bug in which removal of template causes some records
with empty VDSM task ID to be left after command ends.

Change-Id: I97c27cbb8c869182edeefbd3f2cc8c8d466aa29e
Bug-Url: https://bugzilla.redhat.com/990543
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>