core, rest, frontend: fix re-initialization flow
Re-initialize is set by the rest and frontend according to the vm os
type and isInitialized flag in order to do first initialization for the
vm (sysprep/cloud-init).
This logic is not needed (duplicated) at all callers,
as it's required for first initialization always, and all info is in the
db, so this logic moved to runVmCommand.

Now the flag (initializationType) should be used only when caller
(ui,rest) want to run initialization again (re-initialize) or to change
default initialization type, on run once.

Change-Id: I2ff258809c7fedcb7336f4e37be9bc53f2678824
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>