core: Add log for error on run-vm failure
In most cases, when running vm fails there is re-run.
when there is an error in the create process (for example before
createVmVdsCommand is called), the error is not logged anywhere.

This patch adds a log to this error,
also remove the word "VDSM" from the exception message,
unfortunately its not always vdsm fault.

Change-Id: Ia936fee01fa781eff5d5b3959ae6478129c2a136
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>