restapi: del storage conn if domain creation fails
Delete storage connection if storage domain creation fails
in order not to leave leftovers in db from a failed operation.
Note - the connection will be deleted only if it was created as part
of the domain creation call. If it was created in the past and just
reused with its connection id in the domain creation call, then
it will not be deleted.

Change-Id: I71c689ccdbcd038d1b92629b3e46691be0885321
Bug-Url: https://bugzilla.redhat.com/991470
Signed-off-by: Alissa Bonas <abonas@redhat.com>
Signed-off-by: Daniel Erez <derez@redhat.com>