core: call connectStorageServer when add a domain
If a storage domain is created with an already existing storage
connection, it's possible that the storage connection is not mounted
on the host via which the storage domain creation is done, thus
resulting a failure in domain creation.
This patch adds a connectStorageServer call in domain creation flow
to make sure the connection is valid and mount it before the call
to create the StorageDomain.

Change-Id: Ib93e681a36f439277054bf22d8ecb319e330d14f
Bug-Url: https://bugzilla.redhat.com/991797
Signed-off-by: Alissa Bonas <abonas@redhat.com>