core: Accept directory user in AddUser command
Currently the parameter class for the AddUser command has fields for
VdcUser, LdapUser and LdapGroup, but the command only works correctly if
it gets a VdcUser or a LdapGroup, it will fail if only given a LdapUser.
In the future we wan't to be able to use only directory users and groups
(althought groups should go to a different command) and not VdcUser. To
simplify this future change this patch modifies the command so that it
will work correctly if only given the LdapUser.

Change-Id: I5caee3a8e295a13c52ed7ea243b1fc25b7603da6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>