Use FQDN from guest agent for RDP
This commit adds the support for retrieving the guestFQDN value from VDSM
and stores it in the VmDynamic database.
We're not storing FQDN which are localhost or localhost.localdomain

Additionally a new vdc_option "RdpUseFqdnIfAvailable" has been introduced,
to give the administrators the option to disable this feature.
If RdpUseFqdnIfAvailable is true (default) RDP will use the reported FQDN
if it is not null or an empty string.

The value can now be read on the Frontend from the RdpConsoleModel and
it will check whether or not the vdc_option RdpUseFqdnIfAvailable was set.

The FQDN value is now also part of the GuestInfo data in the REST API.

Change-Id: I1e7f8e46c6a71ccbfe9af9bf85271d7294b03526
Bug-Url: https://bugzilla.redhat.com/1014546
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>