core: pre processing for search-generated ldap filters
The search generates a uniform ldap uery in format that contains
upper case keys with $ sign before them - for example
($PRINCIPAL_NAME=a@ovirt.org)

In case of open ldap and RHDS, principal name (i.e - user@domain)
should not be passed to the ldap server as the ldap query
that is being generated uses the "uid" attribute which is based on
user name and not UPN.
Furthermore, at least some of the versions of RHDS, and open ldap do
not contain support for UPN upon out of the box installation.

Instead of passing the UPN to the query, it should be split, and the
ldap query should be run
with a filter that contains only the user part.

Change-Id: I35744eb423ca7f553bf7e9b20c72179bca4e6827
Bug-Url: https://bugzilla.redhat.com/991800
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>