frontend: mouse selection of ListModelTypeAheadListBox
If you are not using the ui binder and select the value
on the ListModelTypeAheadListBox using mouse, the
ListModelTypeAheadListBox.asEditor().getSubEditor().getValue();
does not return the selected value.

The problem was that the mouse selection fired the ValueChangeEvent
but no one was listening on it.

It worked with the editor driver because UiCommonEditorVisitor
listens to this event.

Fixed by listening to this event also in the ListModelTypeAheadListBox.

Change-Id: I5edee9d0b0e64dc4f7e8609aa5fde4c69e133924
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>