core: Remove member init in the compat module
Since the upgrade to GWT 2.5.1, we've seen several cases of GWT pruning
member initialization.

To prevent these (and future) bugs, this series of patches removes all
such initializations from the packages that should be compiled by GWT.
These initialization are either moved to the appropriate
constructor(s), or removed them completely if they are redundant.
Constants with the wrong modifiers are amended to have static final, and
are allowed to be initialized directly.

As a side bonus, this patch offers a minor performance improvement, as
some of these members are now only initialized once instead of twice
(once in the member initialization and then again via a constructor
parameter).

This patch takes care of the compat package, as noted in the subject.

Change-Id: I57bc886e4e53ba7842b435d516604e41b7e469e6
Signed-off-by: Allon Mureinik <amureini@redhat.com>