core: Remove member init in business entities
Since the upgrade to GWT 2.5.1, we've seen several cases of GWT pruning
member initialization.

To prevent these (and future) bugs, this series of patches removes all
such initializations from the packages that should be compiled by GWT.
These initialization are either moved to the appropriate
constructor(s), or removed them completely if they are redundant.
Constants with the wrong modifiers are amended to have static final, and
are allowed to be initialized directly.

As a side bonus, this patch offers a minor performance improvement, as
some of these members are now only initialized once instead of twice
(once in the member initialization and then again via a constructor
parameter).

This patch takes care of all the project's business entities, as
noted in the subject.

Change-Id: I8e8eb9ccb0e68db936489110d7f5c5828d07fab4
Signed-off-by: Allon Mureinik <amureini@redhat.com>