core: Remove member init in the common package
Since the upgrade to GWT 2.5.1, we've seen several cases of GWT pruning
member initialization.

To prevent these (and future) bugs, this series of patches removes all
such initializations from the packages that should be compiled by GWT.
These initialization are either moved to the appropriate
constructor(s), or removed them completely if they are redundant.
Constants with the wrong modifiers are amended to have static final, and
are allowed to be initialized directly.

As a side bonus, this patch offers a minor performance improvement, as
some of these members are now only initialized once instead of twice
(once in the member initialization and then again via a constructor
parameter).

This patch takes care of the class for the common package that were
not handled in previous patches in this series.

Change-Id: Iecdb2abfb2e0a85f0073b0424e7936ff566f770a
Signed-off-by: Allon Mureinik <amureini@redhat.com>