core: Change enum JSON deserialization
This patch introduces an enum deserializer to replace the behavior
that in case of an unnown literal an exception would be thrown.
This may happen in the future, if a Command parameter class with an enum
field is stored in DB, and then oVirt engine is upgraded, and the new code
contains a change in the enum literal.

Change-Id: Ie9d6f2e4a02eeda999e525101c2f54789a34cfaf
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>