engine: Reuse VmInfoBuilder to add profile data to nic in VDS commands
HotPlugNicVDSCommand, UpdateVmInterfaceVDSCommand and VmInfoBuider
has the same logic for adding the profile data to the nic on the struct
sent to vdsm.

This patch adds a static method to VmInfoBuilder- addProfileDataToNic(..).
Using this method also fixes some bugs-
- Makes VmInfoBuilder work with VmNic instead of VmNetworkInterface.
- VmInfoBuilder didn't use the vnic_profile to get the network name and
port mirroring.
- UpdateVmInterfaceVDSCommand didn't add the
vm.getRuntimeDeviceCustomProperties().get(vmDevice) to the custom
properties that were sent to the host.

Change-Id: I234bb2854a5d2d8bdb97dd7d76d34167016c87e7
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>