frontend: handling clusters without DC in VM dialogs
There is a chance that there will be a cluster which does
not have a DC (e.g. the DC has been deleted but the cluster
was not). If this happened the UnitVmModel failed on NPE
for new/edit VM if some DC has been selected in the system
tree.

Fixed by ignoring this clusters in the cluster selection dialog.

Change-Id: Iab60c088d2299b9f7a9153e917c3a9ceaaa0e2fc
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=988018
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>