core: Always use IPv4 for AJP
Currently in the Apache configuration that we use for the connection to
the application server we use "localhost" as the address of the AJP
connector. This host name is usually translated to both the IPv6 ::1
address and the IPv4 127.0.0.1 address, and it looks like Apache always
tries first with the IPv6 address. But the application server doesn't
listen in the IPv6 address, only in the IPv4 address. This means that
the first attempt to connect is always rejected. Not a big issue, as the
web server retries inmediatelly with the IPv4 address, but this
generates an unnecessary roundtrip. This patch replaces "localhost"
with "127.0.0.1" so that the web server will always use IPv4.

Change-Id: I36ef2de20236f509b16b5bb7112d0f849e8e96e1
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>