webadmin: Error when changing email address for user event notification
When changing user event notification email, the
web admin portal complains that the user is already
subscribed to the events.

This happens because we are executing 2 AsyncCalls from
GWT to backend RemoveEventSubscription and AddEventSubscription.
If all RemoveEventSubscription calls do not complete
before AddEventSubscription action is called the exception
is thrown.

Change-Id: If30c1e935cad06b293589c80922f56ee377375cb
Bug-Url: https://bugzilla.redhat.com/985635
Signed-off-by: Ravi Nori <rnori@redhat.com>