core: avoid disconnecting from used connections
When attempting to extend a LUN based storage domain, the engine
attempts to connect all the hosts to the lun. If one or more hosts fail
to see the chosen luns for the extension - the engine attempts to
disconnect all the hosts from the target.

In that scenario the connection isn't loaded from the DB, therefore it's
id will be always null - which can cause to disconnection from targets
that are used by other luns that are in use.

Change-Id: I7008e64ebce57fbbdb350810a5ffca6e0a7800cc
Bug-Url: https://bugzilla.redhat.com/982707
Signed-off-by: Liron Aravot <laravot@redhat.com>