core: Disable member init for GWT-compiled code
Following the upgrade of GWT(P), if an entity's constructor is not
explicitly referenced in the UI code the inline member initializations
are not taken into account in GWT's serialization policy [1].

This patch is the culmination of a series that suggests taking the safe
route - instead of having the code (silently!) break if a call to the
constructor is removed from the UI code, it simply disallows having such
inline initializations.

The previous patches in this set remove such initializers for various
parts of the common and compat modules (broken up into smaller units for
relative ease of review).
This patch implements a checkstyle check to fail the build in compile
time if such a dangerous initializer is introduced.

[1] http://lists.ovirt.org/pipermail/engine-devel/2013-July/005256.html

Change-Id: Iffd520ecb627e87254c4d325c85088179b0ef744
Signed-off-by: Allon Mureinik <amureini@redhat.com>