frontend: Fix for showTooltip condition in AbstractCellWithTooltip
Showing tooltips with the content only works (on Firefox and Chrome
at least) if the content overflows AND contains whitespaces.
If it does not contain whitespaces it will not work. The additional
condition added by this patch will fix this issue by comparing the
actual size of the column with the scrollSize (height and width)

If either of them differs, in the sense that the clientSize is smaller
then the condition is true which means it does indeed overflow.

Change-Id: I26f4942c4f24beedd6b22261f8928bc4b1af5e50
Bug-Url: https://bugzilla.redhat.com/996496
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>