core: make VdcQueryReturnValue#getReturnValue generic
VdcQueryReturnValue#getReturnValue method wasn't generic, so we needed
to have explicit casting which required @SuppressWarnings("unchecked")
annotation which makes the code dirty or add unchecked warning if the
annotation is missing.

This patch changes this method to be generic, and the returned object is
casted to the returned type of the method - this way the annotation
appears only within this method and the rest of the code is cleaner.

In this patch I changed the classes within org.ovirt.engine.core.bll
package to take advantage of this change. Latter patches should change
the rest of the classes, and do the same change in
VDSReturnValue#getReturnValue.

This patch also contains minor refactoring in VdcQueryReturnValue:
1. remove '_' prefix from class fields
2. move the constructor to be before the other class methods

Change-Id: I57b6ce5fc4b5d6bab1481f9bfb9d7861ba09a615
Signed-off-by: Arik Hadas <ahadas@redhat.com>