core: LocalAdminPassword exception after engine restart
After oVirt host comes up from reboot, we see
javax.crypto.BadPaddingException: Data must start with zero
exception in engine.log:

When GetConfigurationValuesQuery is executed it
retrives the values for all the versions and returns
it. But LOCAL_ADMIN_PASSWD exists only for version
general which breaks the logic in DbConfigUtils.

This patch introduces a black list of
all passwd fields that have values only
for version "general". So the backend does
not try to extract values for other versions
for these ConfigurationValues

Change-Id: I000bbb3cf5a6d1d636c49247e85858b02b2f358a
Bug-Url: https://bugzilla.redhat.com/994454
Signed-off-by: Ravi Nori <rnori@redhat.com>