core: scheduling: remove migration domain filter
Currently there is on way to migrate a vm between clusters, therefore
migration domain filter is causing a problem when it's not enabled.
This patch removes the filter, and added it's logic (up state and same cluster)
to the initial host list provided to the filters.

This patch also includes that balancing initial hosts list to be hosts
without migrating vms to allow backward compatability.

Change-Id: I5aea1d3d1e5e63f03ab0d012d4d3fe9356f26e48
Bug-Url: https://bugzilla.redhat.com/988306
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>