engine: Modify vm_interface-vnic-profiles foreign key
The patch modifies the FK of vm_interface on vnic-profiles
to be set NULL in case of deletion of the vnic profiles.

Removal of a vnic profile which is used by vnics is blocked
by can-do-action, same for network removal which is used by
vms.

However, when forcibly removing a data-center, the removal of
the vm interface is being executed in a different transaction
than the network entities.

This can be avoided if the logic of DC removal would have been
implemented by code and not in a stored procedure.

Change-Id: I2feef15a7460899979a78b357886bd1f027918af
Bug-Url: https://bugzilla.redhat.com/996934
Signed-off-by: Moti Asayag <masayag@redhat.com>