core: InitVdsOnUp - flow fixes
1. If the master domain is inactive/unknown - a host shouldn't fail this
phase of the flow if it respondes to getVdsStats.

2. If the master domain is inactive/unknown - the host shouldn't fail
the flow according to it's domain report.

Examples for relevant scenarios:

1. When there are no active hosts, all the domains are marked as
"unknown", if we will attempt to activate a host and it won't be able to
connect to the pool because of VDS_NETWORK_ERROR, the host won't move to
UP because of the check of it's domain report.

2. When there are no active hosts and a host is being activated,
successfully connects to the pool and it can't see any of the domains
it won't move to UP because of it's domain report.

Change-Id: I61fcc58be133efdecf264c22015609f6ee3ed942
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1031632
Signed-off-by: Liron Aravot <laravot@redhat.com>