packaging: setup: allinone: DC version
Before this change, the allinone plugin created a DC with version
being the version of the ovirt-engine package. Currently this fails,
because the engine's version was bumped to 3.4, while the maximum
supported version is 3.3. In principal, we can now add a supported
version '3.4', but there is no reason to do so as long as we do not
have to for compatibility reasons.

This change creates the DC with version being the minimum of the
package version and the maximum supported version.

Change-Id: I01054b74f26fc961633d8f0a55429569dcc8cc31
Signed-off-by: Yedidyah Bar David <didi@redhat.com>