engine: Speed up SPM re-selection process.
Engine trying to send stopSPM command during SPM re-selection proccess.
HSMGetAllTasksStatuses command, which is called as a part of the stopSPM
became the SPM command so technically it can get 'No SPM' error from
VDSM if the SPM is already stopped by other thread. "Not SPM" response
should be treated as kind of success so engine will move forward and
re-select the SPM.

Change-Id: I2c191ad8d4ec0fb4f5f8ff68b84fa5189ecc7114
Bug-Url: https://bugzilla.redhat.com/986961
Signed-off-by: Sergey Gotliv <sgotliv@redhat.com>