core: Force remove storage leaves VM permissions.
Remove of VM will not remove its permissions, since before the remove of
the permissions we remove the image from image_storage_domain_map which
vm_images_view is based on, and therefore it will not fetch this image.

The proposed fix is to create a temporary table which will contains the
vm ids and will not influenced by former operations.

Change-Id: I7445f9ba7c08936e66cab0b043f4dbaaa483e482
Bug-Url: https://bugzilla.redhat.com/999048
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>