core: treat 'special' devices as managed when importing templates
During export, some devices are mapped to "OTHER" resource type in the
ovf. this is valid for unmanaged devices and also some managed devices that are
treated as "special".

When importing VMs, special devices with "OTHER" resource type are
imported as manged.
this patch add the same treatment for import of templates.

Change-Id: I95c1c2f1e2f45a1863c28c5fcaeb6274f2d28b4e
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>