core: simplified RunVmValidator#canRunVm method
The RunVmValidator#canRunVm should be a utility method that invoke the
validations which are required for running a VM, but it was also used to
have logic which determines whether to invoke images related validations
or not.

This patch changes the canRunVm method to invoke all the validations
(until the first failure) without any filtering. Instead, The images
related validations are checking whether the VM has images or not
inside.

Change-Id: I277a24fa26bb9aae3641aa316689ff7350e9d63f
Signed-off-by: Arik Hadas <ahadas@redhat.com>