userportal : Fix DiskForVmGuid high CPU
In this patch we remove the need to load all disk data and just load the
required information for the user portal.

This patch was first committed and merged as
http://gerrit.ovirt.org/#/c/16657/ and then reverted due to DAO tests
failure.
this commit has exactly the same changes as the original patch, the only
change is in the TEST class, the problem was that the test checked the
description field for NULL while it was set by the fixtures.xml to "New
Description"

Change-Id: I1d7369cfbfe9b7c6901bd1bc2e9c16bdc31a9af9
Bug URL: https://bugzilla.redhat.com/show_bug.cgi?id=971237
Signed-off-by: Eli Mesika <emesika@redhat.com>