engine: cleanup AddVdsCommand locking
AddVdsCommand was overriding getExclusiveLocks
to lock the cluster if cluster supports gluster service.

But this is not required here as the "gluster peer probe"
which requires the lock only happens after the host comes
up.

Removed the code which does a "gluster peer probe" in
case ConfigValues.InstallVds = true. This code gets
executed as InitVdsOnUpCommand and is not required
here.

Change-Id: I6189d7f0046a943a58db9d6734c92fe462c9db52
Bug-Url: https://bugzilla.redhat.com/998799
Signed-off-by: Sahina Bose <sabose@redhat.com>