core: Make coverity happy
Some recent changes have introduced a construct that forces the java
compiler to infer the type of a generic method from its use inside the
condition of an if, for example:

  if(Config.GetValue(ConfigValues.ExternalSchedulerEnabled)) {
    ...
  }

The usual (which doesn't mean nice) syntax for this is the following:

  if(Config.<Boolean> GetValue(ConfigValues.ExternalSchedulerEnabled)) {
    ...
  }

While the recently introduced syntax is nicer, unfortunatelly it makes
coverity choke.

This patch reverts to the old style, just to make coverity happy.

Change-Id: I2eb133de6fab92eaaf0edd8a99af2fc406f6b088
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>