packaging: setup: remove dependency of openssh
openssh dependency was left only for getting public key fingerprint.

as we have the logic of doing that anyway because of old rhel server, we
can calc the fingerprint our own.

the ssh-keygen -l is also ugly as it does not support piping the key and
forcing use of temp file... so best manage without it.

Change-Id: Ifa4893d8f9b034642d2428d926d41dd677a23565
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>