core: acquire locks in standard way in MoveOrCopyDiskCommand
In order to solve the race between VM migration and storage commands, we
need the storage commands to be able to acquire VM lock for their entire
execution process. That thing cannot be done when the locks are acquired
using explicit invocation of acquireLockInternal method, we need the
LockIdAndNameAttribute annotation.

Thus, this patch converts the way MoveOrCopyDiskCommand acquire its
locks such that it will be in the standard way.

Change-Id: Id4b5ef206581e5f9a1727c85b30f39802859583c
Bug-Url: https://bugzilla.redhat.com/952147
Signed-off-by: Arik Hadas <ahadas@redhat.com>