webadmin: adding stop rebalance action to volumes tab
Added a new sub-action 'Stop' under the 'Rebalance' action group in the
volumes tab.'Stop' rebalance action will be enabled if the Volume status is UP
and a rebalance operation is already going on.

On clicking 'Stop' rebalance, a confirmation window will be show
to confirming the stop rebalance operation.

Change-Id: I6afc4b3ae9dbea60713178da51700b3197c38bfe
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>