gluster: Added app_mode field to roles table
Added a field app_mode to roles table. The value of application modes
for a role decides if that role is available in that application mode.

This detail would be used for rendering the list of roles as per the
application mode. This would avoid deletion of unwanted roles in case of
gluster downstream.

Change-Id: Id83b7a2c90100744487b276642158e7c3e93de85
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>