frontend: Console reconnect warning dialog changes
This patch adjusts conditions for displaying warning dialog (for
noticing portal when another user is connected to a console). This patch
adds a condition that prevents displaying this dialog when current
portal user has no reconnect permissions. Currently, the warning dialog
would be displayed and after clicking "ok", another dialog with error
would be displayed because of the backend validation. This patch causes
only the "backend" warning to be displayed.

Change-Id: I141d61655e70ab7c0ef89fb7cb1f0c142494bdf8
Bug-Url: https://bugzilla.redhat.com/975729
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>