packaging: setup: firewall-manager - check status
During setup, ask only about firewall managers that are both enabled
and active.

Currently this means that if both firewalld and iptables are enabled,
but iptables is active and firewalld isn't, first we ask about firewalld,
and if users accepts, we do not even ask about iptables. With this
change, at this state, we only ask about iptables.

Change-Id: Ice741b9738c822d48ab3719dd34cbe989b359afe
Signed-off-by: Yedidyah Bar David <didi@redhat.com>