webadmin: avoid disconnect server on attach domain
During the "Import Domain" flow (mostly driven by the UI) the execution
of AttachStorageDomainToPool and DisconnectStorageServerConnection were
involved a race that in some cases resulted in the attach operation to
fail (since the server was disconnected).

Now after attachStorageToDataCenter we avoid postImportFileStorage
(since the storage should remain connected), and we go straight to
onFinish because any issue in AttachStorageDomainToPool is handled by
the default error dialog.

This fix is just removing the race between the calls and it's not
intended to address other bugs that are present in the flow (for example
what happens if the user closes the web browser during the chain of
execution? who is cleaning up the storage connections? etc.)

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=994463
Change-Id: Ie880c25014c5292ad64ed94ad9c126fdd048d7bc
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>