webadmin: hide memory checkbox from create snapshot popup on error
The checkbox that allows the user to select whether to save the memory
as part of the created snapshot or not, was visible also when the create
snapshot popup presented an error message. This patch hides the memory
checkbox from the create snapshot dialog in case it presents an error.

Change-Id: I4b57dc687e19fecbadbec331627f08128b1d6ca0
Bug-Url: https://bugzilla.redhat.com/993645
Signed-off-by: Arik Hadas <ahadas@redhat.com>