engine: Use old method of network init on 3.2 hosts
When using setup networks to set the host's management network instead
of the host deploy, the operation can sometimes fail as super VDSM
service is not ready yet (although VDSM seems active).

We have considered a possibility to retry the operation, but there is no
good way to know what caused the first attempt to fail, or how many
attempts should be made.

Hence, the chosen solution was to have the old way for 3.2 hosts, and
for 3.3 hosts only the setup networks post-installation will occur.

Change-Id: Icd1f41ee037f46a7e9ca360f3d10d5d2a5bcc547
Bug-Url: https://bugzilla.redhat.com/999461
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>