core: Introducing VersionStorageFormatUtil
Introduced a new utility class, VersionStorageFormatUtil, to handle the
relationship between Version and StorageFormatType.

This patch includes the following:
1. The new VersionStorageFormatUtil class.
2. A test case for this class.
3. Refactoring UpdateStoragePoolCommand to use this new class.
4. Refactoring AddStoragePoolWithStoragesCommand to use this new class.

Note that a side benefit of changes (3) and (4) is future-proofing
these commands in a way that if a future version will introduce
storage format V4, these classes won't have to be touched.

Change-Id: I47fb8b14e5c97b601c8fc0cc296a4ec42a48227a
Signed-off-by: Allon Mureinik <amureini@redhat.com>