frontend: incorrect DC/Cluster handling on new pool
Have one DC with two clusters. Make a template on the first cluster.
Go to the new pool dialog and select the second cluster - it gets
automatically reset to the first one.

The problem was that the logic of selecting the DC/Cluster was
implemented in the common place for new and edit dialog. For edit
dialog it selected the DC/Cluster according to the selected pool.
For new pool according to the selected template.

While for edit pool it made sense to select the DC/Cluster according
to edited pool it made no sense for new pool to select the DC/Cluster
according to the selected template.

Fixed by moving the logic specific for edit pool to the
ExistingPoolModelBehavior from PoolModelBehaviorBase.

Change-Id: I38e771d88fc300304942f750b6628862279a02ad
Bug-Url: https://bugzilla.redhat.com/1002902
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>