engine: VM NewPool-1 was restarted on Host <UNKNOWN>
If the host has been started as prestarted the following
event log has been logged:
VM NewPool-1 was restarted on Host <UNKNOWN>

The problem was that the Vm has been started as stateless.
This first triggers the statelessVmTreatment() which
does not schedule the host but will already be logged (e.g. the host will be
unknown).

Fixed by checking if the host has already been scheduled and
if not than a different message will be logged (a similar to the one when
the VM is started as stateless by the user).

Change-Id: I60a74762984dc82154d0f8e073c60f631da229f0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=966619
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>